Hybrid-pinecone/requirements.txt
pinecone-client
sentence-transformers
gradio
pymysql
pandas
